+++ NEWS.md
@@ -2,10 +2,13 @@
 # hclust1d 0.0.1.9006
 
 - Added a vignette on replacing `stats::hclust` with `hclust1d`
+- Added citation info
+- updated `.onAttach` with citation info
 
 # hclust1d 0.0.1.9005
 
-- Added `ward.D`, `ward.D2` linkage support and a getting started vignette
+- Added `ward.D`, `ward.D2` linkage support
+- Added a getting started vignette
 
 # hclust1d 0.0.1.9004